requirements-mac.txt
# Mac M4 Pro optimized requirements for Fountain Buddy
# Python 3.13 compatible version - Use this instead of requirements.txt on Apple Silicon Macs

# Core dependencies
requests>=2.31.0
numpy>=1.24.0

# Computer Vision - Apple Silicon optimized
opencv-python>=4.8.0
pillow>=10.0.0

# AI/ML - Apple Silicon TensorFlow (Python 3.12 compatible)
tensorflow-macos>=2.15.0
tensorflow-metal>=1.1.0
tensorflow-hub>=0.15.0

# YOLO - Ultralytics with Apple Silicon support
ultralytics>=8.0.0

# Discord and scheduling
discord.py>=2.3.0
apscheduler>=3.10.0

# ML utilities
scikit-learn>=1.3.0

# Environment
python-dotenv>=1.0.0

# Python 3.12 compatibility fixes
setuptools>=69.0.0